Omni translation-unit topology
MxOmniCreateFlags and MxOmniCreateParam fold into mxmain.cpp,
MxScheduler folds into mxthread.cpp, and the stream providers move into
their own streamprovider library per the retail link layout. DllMain
came from the CRT, so main.cpp gives way to the library_msvcrt.h
annotation header, and define.cpp joins the omni library.
