+++ LEGO1/library_msvcrt.h
@@ -1,0 +1,2 @@
+// LIBRARY: LEGO1 0x10091ee0
+// _DllMain@12
