--- LEGO1/main.cpp
@@ -1,7 +1,0 @@
-#include <windows.h>
-
-// FUNCTION: LEGO1 0x10091ee0
-BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
-{
-	return TRUE;
-}
