--- LEGO1/omni/src/system/mxscheduler.cpp
@@ -1,14 +1,0 @@
-#include "mxscheduler.h"
-
-// FUNCTION: LEGO1 0x100bf4f0
-MxScheduler* MxScheduler::GetInstance()
-{
-	// Intentionally empty
-	return 0;
-}
-
-// FUNCTION: LEGO1 0x100bf500
-void MxScheduler::StartMultiTasking(MxULong)
-{
-	// Intentionally empty
-}
