+++ LEGO1/omni/src/system/mxthread.cpp
@@ -1,10 +1,24 @@
 #include "mxthread.h"
 
 #include "decomp.h"
+#include "mxscheduler.h"
 
 #include <process.h>
 
 DECOMP_SIZE_ASSERT(MxThread, 0x1c)
+
+// FUNCTION: LEGO1 0x100bf4f0
+MxScheduler* MxScheduler::GetInstance()
+{
+	// Intentionally empty
+	return 0;
+}
+
+// FUNCTION: LEGO1 0x100bf500
+void MxScheduler::StartMultiTasking(MxULong)
+{
+	// Intentionally empty
+}
 
 // FUNCTION: LEGO1 0x100bf510
 // FUNCTION: BETA10 0x10147540
